POLLY-ONLY: add &operator<<(llvm::raw_ostream &OS, Obj o)
The raw_ostream streaming operator allows to print the isl objects to
the LLVM ostreams. This change is only useful when using isl in the
context of LLVM. It is included in this branch to facilitate the testing
of the isl CPP bindings, but is not intended to be upstreamed to isl.